backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.1.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
# GeoDjango для работы с картами и геолокацией
GDAL>=3.4.0